pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "niwa-qmra-toolkit"
version = "1.0.0"
description = "Quantitative Microbial Risk Assessment Toolkit - Python replacement for @Risk Excel functionality"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "NIWA Earth Sciences", email = "reza.moghaddam@niwa.co.nz"}
]
maintainers = [
    {name = "Reza Moghaddam", email = "reza.moghaddam@niwa.co.nz"},
    {name = "David Wood", email = "david.wood@niwa.co.nz"}
]
keywords = ["microbiology", "risk assessment", "qmra", "niwa", "water quality", "pathogens"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
]
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.21.0",
    "scipy>=1.7.0",
    "pandas>=1.3.0",
    "matplotlib>=3.4.0",
    "python-docx>=0.8.11",
    "PyYAML>=6.0",
    "click>=8.0.0",
    "openpyxl>=3.0.9",
    "seaborn>=0.11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.10",
    "black>=21.0",
    "flake8>=3.8",
    "mypy>=0.812",
    "sphinx>=4.0",
    "sphinx-rtd-theme>=1.0",
]
gui = [
    "tkinter",
]

[project.scripts]
qmra = "qmra_toolkit.src.qmra_toolkit:cli"
qmra-gui = "qmra_toolkit.src.qmra_gui:main"
qmra-enhanced-gui = "qmra_toolkit.src.enhanced_qmra_gui:main"

[project.urls]
Homepage = "https://github.com/niwa/qmra-toolkit"
Documentation = "https://niwa-qmra-toolkit.readthedocs.io/"
Repository = "https://github.com/niwa/qmra-toolkit"
"Bug Tracker" = "https://github.com/niwa/qmra-toolkit/issues"

[tool.setuptools.packages.find]
where = ["qmra_toolkit"]

[tool.setuptools.package-data]
"qmra_toolkit" = [
    "data/*.json",
    "data/*.yaml",
    "config/*.yaml",
    "templates/*.docx",
    "assets/*",
]

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["qmra_toolkit/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--tb=short",
    "--strict-markers",
    "--disable-warnings",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]